+++ src/commonMain/kotlin/ehn/techiop/hcert/kotlin/rules/BusinessRule.kt
@@ -1,8 +1,7 @@
-package ehn.techiop.hcert.kotlin.trust
+package ehn.techiop.hcert.kotlin.rules
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-
 
 /**
  * One specific business rule, as raw JSON
